+++ model/set_bc1.f90
@@ -25,7 +25,7 @@ module set_bc1_module
 
       implicit none
 
-      integer     , intent(in   ) :: slo(3),shi(3)
+      integer(c_int), intent(in ) :: slo(3),shi(3)
 
       real(c_real), intent(inout) :: p_g&
          (slo(1):shi(1),slo(2):shi(2),slo(3):shi(3))
